Allow filtering CiStepResults by branch, passed, & created_at [CSR-3] (#6157)
Ransack docs:
https://activerecord-hackery.github.io/ransack/getting-started/simple-mode/

This will be useful because different CI runs execute different tasks
and in different orders, etc. By allowing filtering, it should be
somewhat possible to view a subset of results that are more closely
related to each other. One especially useful filter will be `branch`,
particularly looking at the `main` branch (which always runs the same CI
steps), so this change filters by the `main` branch by default. This
change also adds the ability to filter by `passed` status (default:
true) and by `created_at` time (default: past 2 months).

This filtering applies both to the "Run times" and "Parallelism" line
charts as well as the Gantt charts.

This initial implementation is quite quick-and-dirty, with all three of
the filtering inputs currently just being text fields. Ideally `branch`
would probably be some sort of typeahead, `passed` would probably be a
select dropdown, and `created_at` would be a date-and-time-picker.
However, in order to get something merged and to keep the change as
small as possible, for now, I think that this is good enough. I have
added a ticket to the backlog to improve the UX of these form inputs:
https://davidrunger.atlassian.net/browse/CSR-9 .