[ci-step-results] Add ability to filter by name/type [CSR-10] (#6167)
The run times line graph doesn't give us the ability to show only a
single type of CI step (except by toggling off every one of the many
other CI step types), which is something that I sometimes want to do
(e.g. to see if RuboCop has slowed down). This makes it possible by
filtering the data on the server side.

(Note: some steps (such as `RubRubocop`) don't run on `main`, which is a
filter applied by default, so one must remember to clear the `main`
branch filter in order to see results for such a CI step that doesn't
run on `main`.)


![image](https://github.com/user-attachments/assets/58b5d9cc-8e95-429b-b677-b14592ead556)