[ci-step-results] Make 'branch' input a typeahead (datalist) [CSR-13] (#6170)
I wasn't sure where to put the `branch_names` method. I stuck it into
the presenter, but this required adding two new initialization params
and also (relatedly) isn't really aligned with the single purpose of
"presenting CI step results". But that's okay for now.

This change uses an HTML `datalist` element to create basically a
typeahead experience using a native HTML text input.

Also, since it was helpful to me in working on this feature, this change
also adds `controllers` and `presenters` to the list of files upon
modification of which Vite will reload the page.


![image](https://github.com/user-attachments/assets/32163bb0-3c57-46bf-b5b5-3cc1c5022dcc)