[blog] Add 'js' to allowlisted extensions [BLOG-17]
As of https://github.com/davidrunger/blog/pull/518 , we are now trying
to serve JavaScript for the blog (to add "Copy Code" buttons to all
`pre` tags). However, it's not working, because we weren't allowing
requests for `.js` files. This change adds `js` to the list of
allowlisted extensions and also disables forgery protection for the
blog#assets action, which will allow that JavaScript code to load
successfully.

Without disabling forgery protection, we get this error when requesting
the JavaScript file:

> ActionController::InvalidCrossOriginRequest: Security warning: an embedded <script> tag on another site requested protected JavaScript. If you know what you're doing, go ahead and disable forgery protection on this action to permit cross-origin JavaScript embedding.