[check-ins] Import ElPopover in component (not entrypoint)
This is largely for consistency with #6135 . In general, my goal is to
have each component that uses a component library's component(s) import
that component library's component(s) in my component(s) that use
it/them. I think that this makes it a little more clear "where things
are coming from" and might also make it easier to replace.