[check-ins] Import ElPopover in component (not entrypoint) (#6187)
The motivation for this change is partially for consistency with #6135 ;
I would have made this change in that PR, if I'd have noticed this
issue, then.

In general, my goal is to have each component that uses a component
library's component(s) import that component library's component(s) in
my component(s) that use it/them. I think that this makes it a little
more clear "where things are coming from" and might also make it easier
to replace. It also provides better TypeScript support. (Indeed, this
change surfaced a TypeScript violation, which was somewhat of the
motivation for the feature improvement herein of linking to the
emotional need edit page for an emotional need without a description.)