[groceries] Auto-linkify URLs in grocery items [GROC-25] (#6195)
Sometimes I include URLs in my grocery items. It will be convenient for
these URLs to be hyperlinked. This change implements that.

I tried to find a micro-library that would do this for me, but I
couldn't find something with the characteristics that I want, so I am
implementing it myself. Hopefully this regex will cover the relevant
(edge) cases as I want it to.

Also, I am adding `"moduleDetection": "force"` to `tsconfig.json` to
avoid the error ("Cannot redeclare block-scoped variable") that I was
getting when editing a file (`app/javascript/typescript-scratchpad.ts`)
that did not have any imports or exports and which had a constant name
(`regex`) that was also in another file
(`app/javascript/lib/linkify.ts`). Source:
https://stackoverflow.com/a/77005521/4009384 .