Add bin script to run datamigrations on the server
When running a datamigration via `docker compose exec`, the output won't
go to our logs (Docker -> Vector -> Loki -> Grafana) by default, just to
stdout. We can use `tee` to both print to stdout and to write to the
stdout of process 1, which will cause the logs to be captured by the
Docker -> Vector -> Loki -> Grafana pipeline. This change adds a script
which does that.

Also, modify `Datamigration::Base` so that it chooses a different
`logger_base`, depending on the environment. See code comments on that
code change for more detail. Without this change, there is not output
when running migrations in development (whether via this new `bin/`
script or via `rails runner` directly).