Add bin script to run datamigrations on the server (#6198)
When naively running a datamigration via `docker compose exec` (e.g.
`docker compose exec web bin/rails runner
db/datamigrate/20250220020837_destroy_orphaned_log_entry_data.rb`), the
output won't go to our logs (Docker -> Vector -> Loki -> Grafana) by
default; it only prints to the terminal.

However, we can use `tee` to both print to stdout and to write to the
stdout (file descriptor 1) of process 1, which will cause the logs to be
captured by the Docker -> Vector -> Loki -> Grafana pipeline. This
change adds a script which does that.

Also, modify `Datamigration::Base` so that it chooses a different
`logger_base` depending on the environment. See code comments on that
code change for a little more detail. Without this change, there is not
output to stdout when running migrations in development (whether via
this new `bin/` script or via `rails runner` directly); instead I think
that the output is written to `log/development.log` (in the container, I
guess).