[groceries] Put the delete-item "X" in a proper button [GROC-26] (#6205)
Currently (on `main`), the click target is pretty small, and there is no
hover effect to indicate whether a click will succeed. This change
improves the UX by putting the "X" to delete a grocery item in a larger
target button with a visible border that will change its appearance
if/when the user hovers over it.

It also changes the "X" from a Unicode symbol to an icon, which creates
better alignment within the button (particularly vertically), similar to
PR #5930.

This change also removes a number of CSS classes that I don't think are
needed / doing anything anymore.