[run-release-tasks] Don't require 'web' service to be running
`docker compose exec web` will error if the 'web' service isn't running,
such as when booting up freshly on a brand new server, or maybe after
there has been some sort of problem. It's not essential that we have an
`OUTGOING_GIT_SHA` (and we explicitly handle the possibility of its
absence where it is checked in `bin/server/release-tasks`), so, rather
than erroring if `web` is not running, this change makes it so that we
will simply set `OUTGOING_GIT_SHA` to noting, in that case.

I noticed this issue while trying to boot up on a fresh Hetzner server
via `GIT_REV=$(git log main --format=format:%H | head -1)
bin/server/deploy.sh`.