[roll-out-web] Don't assume nginx and/or web were booted before (#6207)
The implicit assumptions in `bin/server/roll-out-web.sh` that nginx
and/or web will have been booted before that script is executed cause
`GIT_REV=$(git log main --format=format:%H | head -1)
bin/server/deploy.sh` to fail on `main` when run on a fresh server
(which I discovered when running it on a fresh Hetzner server just
recently).

This change modifies the `bin/server/roll-out-web.sh` script so that it
will work even if `nginx` and/or `web` were _not_ booted before that
script is executed.