[deploy.sh] Behave differently if LOCAL_TEST=1 [DEV-180]
Currently, to test what a deployment will do, I either issue a big
command in my terminal like:

```
{ docker stop $(docker ps -q) || true } && bin/build-docker development && docker compose run --rm web bin/server/release-tasks && bin/server/boot-services.sh && bin/server/verify-expected-services.sh
```

or else copy `bin/server/deploy.sh` to `personal/bash.sh` and tweak it.

This change tweaks `bin/server/deploy.sh` so that it will do what we
want (i.e. skip certain steps and build with the development Rails env)
if run with `LOCAL_TEST=1`. This way, I don't need to run a big command
in my terminal or tweak a copied version of `deploy.sh` every time that
I want to test what a deployment will do.

We add a check at the beginning such that the script will exit with an
error if LOCAL_TEST is _not_ set but if `git config user.email` is
present (i.e. if we are probably on a developer's local machine). This
should decrease the likelihood of running the production deployment
locally.