Log stderr (just not stdout) when precompiling assets (#6214)
The main thing is that we wanted to suppress the verbose output written
to stdout. However, if an error occurs when precompiling assets (which
is actually mostly downloading assets from S3), then it would be very
helpful for that to be logged, which this change will at least make more
likely.