[gitignore] Remove database dump files
For one thing, `bin/qr` only creates files in `tmp/`, which is itself
gitignored, so ignoring these file names is somewhat redundant. That
was, indeed, part of the reason to add these, so that we would have
redundant protection against accidentally committing a database dump
file to git. However, instead of having this layer of redundancy, I am
instead going to lean on a new linter that I plan to create soon that
will fail if any non-image, non-text files are committed to the repo
(https://davidrunger.atlassian.net/browse/DOT-93).