[ci] Wait for Percy to start within StartPercy task itself (#6223)
I think that it used to be the case that RunApiControllerTests and
RunUnitTests completed before JavaScript compilation. When this was the
case, I think it made sense to have a separate WaitForPercyStart task
that ran after those steps.

However, in recent runs, I think that JavaScript compilation has
actually sometimes been completing before RunUnitTests. In this case,
having WaitForPercyStart wait for RunUnitTests effectively delays the
execution of feature specs, and I think thereby slows down the overall
test suite.

Below is the most recent Gantt chart on `main`, illustrating this. We
wait for RunUnitTests (pink) and WaitForPercyStart (red) before feature
specs start executing. During this time, parallelism is low (wastefully
low, I think), with just 2 tasks running.


![image](https://github.com/user-attachments/assets/38e590f8-d0ec-438d-a0a4-f4307c08edb1)

This change moves the checks for Percy startup into the StartPercy task
itself. This way, we will be ready to start feature specs sooner and we
won't need to wait for RunUnitTests + WaitForPercyStart anymore.

This change adds `sleep(2)` to each iteration of the loop, to hopefully
waste a little bit less energy wastefully pinging Percy while it boots
up, now that we will try to check for Percy startup immediately after
launching the Percy server in the background.