Sleep half a second after disconnecting ActiveRecord connections
We saw this error when trying to create the `_unit` database:

```
2025-02-22T10:43:20.3032090Z Running system command '[1;33mcreatedb -T david_runger_test david_runger_test_unit -U postgres -h 127.0.0.1 --no-password[0m' with ENV vars [1;33m{}[0m ...
2025-02-22T10:43:22.8969138Z [percy] Percy is not running
2025-02-22T10:43:23.2371505Z [percy] Successfully downloaded Chromium 1300309
2025-02-22T10:43:23.8410170Z [percy] Percy has started!
2025-02-22T10:43:25.4655387Z createdb: error: database creation failed: ERROR:  source database "david_runger_test" is being accessed by other users
2025-02-22T10:43:25.4656690Z DETAIL:  There are 2 other sessions using the database.
2025-02-22T10:43:25.4688068Z
2025-02-22T10:43:25.4690866Z [1;31m'[1;31mcreatedb -T david_runger_test david_runger_test_unit -U postgres -h 127.0.0.1 --no-password[0m' with ENV vars [1;31m{}[0m failed (exited with 1, took 5.169).[0m
```

Creating the other four databases succeeded.

This seems pretty suspiciously timed with the database-related change
being made in this PR, but it's not clear to me what the specific
connection could be? It seems like maybe it's just a coincidence?

The fact that the DB creation that errored was for `_unit` -- the first
in the list of DBs to create -- makes me wonder if maybe the
disconnections initiated just prior maybe are asynchronous and hadn't
completed yet?

So, this commit adds a 0.5 second sleep, with the goal of giving enough
time that those disconnections will complete.