Move SECRET_KEY_BASE env var from ruby.yml to .env.test
Also, don't make it look like a real secret.

Without this change, I run into an exception [here][1] when trying to
run tests locally.

[1]: https://github.com/davidrunger/david_runger/blob/681466de66faa03363a7cfe0a72bdcd1f1e4e05b/config/initializers/0w_secret_key_base.rb/#L11

This change is not related to the main purpose of this PR, apart from
the fact that I want to run tests locally to get them passing for this
PR.

This is a followup to #6213 .