[ci] Create a dedicated database for RunFeatureTestsA (#6226)
This way, RunFeatureTestsA won't have to wait for RunUnitTests to
complete before RunFeatureTestsA can start running. Having to wait used
not to be a problem, because RunUnitTests always finished before
CompileUserJavaScript, anyway, but that's not the case anymore.

The goal is to avoid runs like the one below, where RunFeatureTests A
waits for RunUnitTests even though CompileUserJavaScript had already
completed:


![image](https://github.com/user-attachments/assets/f7410562-ad55-4dd1-932e-66ac3568394f)

That run was particularly unlucky because RunFeatureTestsA happened to
be the slowest of the three feature test runners, and it started last,
due to waiting for the unit tests. With this change, there will no
longer be any waiting for unit tests.