Increase Pallets concurrency from 5 to 6 (#6228)
There are some times when we are bound by the concurrency limit of 5 and
would to better to be able to run 6 tasks simultaneously. This is
particularly due to #6223, which causes the `StartPercy` task to take
much longer. Since I think that task largely uses IO (waiting on network
calls to the Percy server and/or sleeping), it's probably not using the
CPU very much when it runs, so I think that (at least during that time)
adding another concurrent task won't add much to the CPU burden.

This change increases the `sleep` in `StartPercy` iterations from 2
seconds to 4 seconds to try to make the CPU burden of `StartPercy` even
lower. The only thing that really matters is that it completes before
`CompileUserJavaScript` finishes, which I think that it still will do.