Create a backup DB in 'bin/qr' and restore from it with 'bin/dbrest' (#6233)
This can be super helpful when repeatedly testing something that
nontrivially modifies the database, such as a datamigration, or even
just a user flow in the web UX that modifies some data.

I think that both the create-backup-DB functionality in `bin/qr` and the
restore-from-backup-DB functionality in `bin/dbrest` require that
nothing be connected to the relevant database(s) when those operations
are performed, so this change adds a small
`bin/terminate-db-connections` script that is shared by both `bin/qr`
and `bin/dbrest` for this purpose. Previously, when I have done such a
thing (at CommonLit), I think that it was necessary to monkeypatch the
Postgres ActiveRecord adapter to try to re-establish a connection, if
the connection found itself to be disconnected (due to having been
disconnected for the sack of restoring from the backup DB, for example),
but it seems that now Rails does this for us? At least, in my testing,
no such monkeypatch seems to be necessary, in the version of Rails that
I am currently on (8.0.1). This just works, as is.