Enable host verification for production (#6240)
I was not able to successfully deploy this when I tried this earlier.
However, I think that has to do with the fact that I tried to deploy it
in two steps (first uncommenting `config.hosts` and then in a later PR
uncommenting `config.host_authorization`), and I think that this wasn't
playing well with how we roll out new web instances to production. I
think it will work if we make both changes simultaneously, which is what
this change does.