Fix DiffHelpers#diff method (#6242)
#6241 should have run RuboCop, but it
[didn't](https://github.com/davidrunger/david_runger/actions/runs/13484547862/job/37674025712).
That's because the DiffHelpers#diff method was broken in #6237 by using
`main-branch`, which is a command that I have on my machine, but which
is not on the GitHub Actions runners.

This change fixes the command by simply hardcoding `main` instead of
using the (nonexistent on GitHub Actions runners) `main-branch` command.

RuboCop has been flaking/failing a lot lately, so this change disables
it for now. Hopefully the bugs will be fixed, soon, although I fear that
they might not be, since I haven't been seeing others reporting these
bugs at the rubocop issues tracker.