[ci] Run unit tests only after API controller tests have completed (#6246)
My thinking is that this will hopefully reduce the parallel CPU demands
during JavaScript compilation, allowing feature tests to start sooner.
It's not clear whether the implications of this will be an overall win
for test suite run time, but I think it might be, and I think it might
be worth giving this a shot.