[ci] Run API controller tests after StartPercy completes
This change builds on #6246 to further try to reduce concurrency while
JavaScript assets are being compiled. My hope is that this will make
JavaScript compilation faster, allowing feature tests to start sooner,
and that this might be worth the price paid in terms of higher
concurrency during feature specs and a later start for unit tests and
API controller tests. It's not clear whether the implications of this
will be an overall win for test suite run time, but I think it might be,
and I think it might be worth giving this a shot.