Set up Blazer [ADM-4]
https://github.com/ankane/blazer

This will be a lot more flexible than our current solution, which is
basically a single hardcoded graph of home#index response times over
the past 24 hours. With Blazer, new queries (and variations of
queries) won't need to be hard-code into the app code.

The Blazer read-only Postgres user needs to be created via `ruby
script/create_blazer_user.rb` or (in Docker) `docker compose exec web
ruby script/create_blazer_user.rb`. This introduces our first file in
`script/`, which is a new standard Rails directory.
https://guides.rubyonrails.org/getting_started.html#directory-structure

> `script/`: Contains one-off or general purpose scripts and benchmarks.

I don't love that the creation of this Blazer Postgres user isn't fully
automated, but unfortunately I think that doing that would be difficult
and is probably not worth it.