[routes] Mount engines using :at keyword
I tried adding a route for local testing purposes like
`get('/blazer-db-url', to: ->(_env) {
plain_text_response(ENV.fetch('BLAZER_DATABASE_URL')) })`. This route
could not be accessed when mounting the Blazer engine via `mount
Blazer::Engine => '/blazer'`; instead, it seems that the Blazer engine
was matching the path but then catching it as a 404, since the Blazer
engine doesn't provide such a route.

In contrast, with the change made in this PR, where we instead mount the
Blazer engine (and similarly the Flipper and Sidekiq engines) using the
`:at` keyword (`mount Blazer::Engine, at: 'blazer'`), such a route can
be accessed. This seems much better to me, and so this change does that.