Don't try to fully destroy a comment that has children
It violates a foreign key constraint, which is the technical
representation of the product reality that we probably shouldn't allow
totally deleting comments that have children/replies, because that would
effectively make the children totally irrelevant / not displayed.
Instead of totally destroying it, if a Comment has one or more children,
we will "blank it out".