Add ability to comment on blog posts [COM-1] (#6292)
This ended up taking a lot more than I expected. The comments work was a
bit more involved than I expected, and also some other changes
scope-creeped into this.

One of the biggest scope-creep changes was some changes to the
`Intersection` type. These changes make that type _far_ more complex
than it was before. I haven't even reviewed that in detail; ChatGPT
wrote it for me:
https://chatgpt.com/c/67c03a03-cc4c-8007-9962-6c22ed3bb313 . But the
upshot is that I think it works better than it did before, in a few
ways. For one thing, it now works for array input types, whereas before
it basically didn't meaningfully work for them (apart from effectively
making sure that both elements are arrays, I think).

Another notable scope-creep change is the addition of `decorate_with`
functionality to our serializers. Unfortunately, this involved some
light monkeypatching of Alba. But I think it will be good to be able to
use decorators in serializers. I'm a little surprised that I haven't
introduced this already.

I also added a good amount of logic related to redirecting users through
a wizard flow after they log in. This code is kind of ugly/hacky, and
could probably do well to be extracted to a concern and/or to be
abstracted into some class(es) or something. For now, though, it serves
the purpose (giving new users the opportunity to choose a display name),
and so I am just going to go with it.