Dedupe dependencies in pnpm-lock.yaml
`pnpm dedupe`

For some reason (presumably some sort of bug in `vega-expression`?), I
think that this causes some(/all?) of the `vega-expression` code to be
included twice, which significantly increases the size of the
`ci_step_gantt_charts*.js` package. Whatever. Hopefully they will fix it
at some point (if, indeed, it's really a bug; I think it probably is).
But, since this bundle is not really used by users other than me, I'm
not going to worry about it.