Remove app-wide registrations of Modal component [DEV-197] (#6305)
**Motivation:** I think that this will give us better syntax
highlighting and also probably better TypeScript support.

I would like to be able to enable the `vue/no-undef-components` rule to
detect this issue programmatically, but it doesn't work with pug
templates. There is this
https://github.com/rashfael/eslint-plugin-vue-pug , which probably could
make it work with pug templates, but I don't love the idea of adding a
quasi-random package like that, and also that package isn't yet natively
compatible with eslint's new "flat config"
https://github.com/rashfael/eslint-plugin-vue-pug/issues/
29#issuecomment-2459341391 (and there is no reply from the maintainer to
a question asking about it). So, I will just rely on myself, rather than
tooling.