Use 'tsd' for 'Intersection' type testing
It occurred to me that we weren't doing any testing of negative
expectations for `Intersection`, and that we cannot do such testing just
using TypeScript types and checking that the compilation is successful.
But maybe the tests are passing because `Intersection` just returns
`any`. Using `tsd` allows us to also make negative assertions, which
allows us to check that the `Intersection` is not overly _permissive_,
in addition to the assertions/tests/examples we already had checking
that it is not overly _restrictive_.