Suppress stderr in CompileAdminJavaScript
There is a lot of output about `DEPRECATION WARNING [global-builtin]:
Global built-in functions are deprecated and will be removed in Dart
Sass 3.0.0.` and other such stuff, which I believe originate in the
ActiveAdmin Sass. Previously, when running via `vite_ruby`, I guess that
this went to stdout, but now it seems to go to stderr.