[CreateDbCopies] Remove disconnect retry and debugging query
I'm optimisit these won't be needed anymore and that the change in this
PR will resolve the issue that prompted the addition of these things. If
this proves not to be the case, we can add this stuff back. (I don't
think that the 50 attempts to disconnect loop was really doing anything
substantive/helpful, though, so maybe don't add that back.)