[ci] Don't have RunApiControllerTests wait for StartPercy [DEV-200]
Compiling JavaScript is much faster now because of #6318  .

Some of the dependency relationships in `RequirementsResolver` were
predicated on the assumption that compiling JavaScript would take a
certain length of time, and I think that some of those assumptions no
longer hold.

One such assumption is that, even if RunApiControllerTests starts after
StartPercy, then it will still complete before JS compilation is
finished. This is no longer true, and it results in unnecessary
downstream delays in the execution of CI steps, with ultimately
RunUnitTests sometimes being the last CI step to finish.

Accordingly, this change removes the dependency of RunApiControllerTests
on StartPercy, which will allow RunApiControllerTests to start running
earlier, which will hopefully produce an overall faster and more
efficient build timeline.