Remove RunUnitTests => RunApiControllerTests dependency [DEV-200]
I think that this made more sense back when JS compilation took longer,
and so RunApiControllerTests and then RunUnitTests could complete in
series in not much more time than it took to compile the JavaScript.

However, now that JavaScript is compiling much faster, when RunUnitTests
depends on RunApiControllerTests, RunUnitTests ends up running mostly
after the compilation of JavaScript, in parallel with the 3 feature
tests runners and HTML tests. I think that this backloads too much
concurrency and we will probably get faster builds by starting
RunUnitTests earlier, which this change accomplishes by removing the
artificial dependency on RunApiControllerTests.