[spec/features/logs] Add 200 ms sleep before downloading CSV [LOG-38] (#6358)
Hopefully this will fix some flakiness that we have seen in this spec
(both in CI and locally).

On my machine, this was flaking about 3% of the time before this change.
With this change, I ran the spec 200 times and didn't get a failure.

The feature spec server logs added in #6003 were really helpful, because
they showed that no request to download the logs is reaching the server
when the spec fails, which focused the possible causes of the flakiness
and the possible fixes to try.