Replace rspec-instafail formatter with built-in failures formatter (#6363)
Don't we nowadays [buffer all of the
stdout](https://github.com/davidrunger/david_runger/blob/1d380df8d824d8e8f137c1f4e26274a2c9611a42/lib/test/task_helpers.rb#L23-L30)
until each rspec test suite completes, anyway, thereby defeating the
whole purpose of this formatter? The result is just that each failure
gets printed twice, which is not very helpful.

This change replaces the rspec-instafail formatter with the built-in
failure formatter. This way, when there is a failure, concise info about
the failure is printed (file, line, description), but no more. The
advantage of printing a little info is that it makes it clear which spec
the failure screenshot(s) are for.


![image](https://github.com/user-attachments/assets/88b431fc-e692-40c5-8c77-78dbdea3c9d9)